+++ binance_websocket_api_futures.py
@@ -1,37 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-#
-# File: binance_websocket_api_futures.py
-#
-# Part of ‘UNICORN Binance WebSocket API’
-# Project website: https://www.lucit.tech/unicorn-binance-websocket-api.html
-# Github: https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api
-# Documentation: https://unicorn-binance-websocket-api.docs.lucit.tech
-# PyPI: https://pypi.org/project/unicorn-binance-websocket-api
-# LUCIT Online Shop: https://shop.lucit.services/software
-#
-# Author: LUCIT Systems and Development
-#
-# Copyright (c) 2019-2024, LUCIT Systems and Development (https://www.lucit.tech)
-# All rights reserved.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a
-# copy of this software and associated documentation files (the
-# "Software"), to deal in the Software without restriction, including
-# without limitation the rights to use, copy, modify, merge, publish, dis-
-# tribute, sublicense, and/or sell copies of the Software, and to permit
-# persons to whom the Software is furnished to do so, subject to the fol-
-# lowing conditions:
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-# OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABIL-
-# ITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT
-# SHALL THE AUTHOR BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
-# WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-# IN THE SOFTWARE.
+# ¯\_(ツ)_/¯
 
-from unicorn_binance_websocket_api.manager import BinanceWebSocketApiManager
+from unicorn_binance_websocket_api import BinanceWebSocketApiManager
 import asyncio
 import logging
 import os
@@ -73,7 +44,6 @@ async def binance_stream(ubwa):
     ubwa.stop_manager()
 
 if __name__ == "__main__":
-    logging.getLogger("unicorn_binance_websocket_api")
     logging.basicConfig(level=logging.DEBUG,
                         filename=os.path.basename(__file__) + '.log',
                         format="{asctime} [{levelname:8}] {process} {thread} {module}: {message}",
@@ -81,7 +51,7 @@ if __name__ == "__main__":
 
     # To use this library you need a valid UNICORN Binance Suite License:
     # https://shop.lucit.services
-    ubwa = BinanceWebSocketApiManager(exchange='binance.com')
+    ubwa = BinanceWebSocketApiManager(exchange='binance.com-futures')
     try:
         asyncio.run(binance_stream(ubwa))
     except KeyboardInterrupt: